+++ lib/src/ext/scoped_widget_state.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:simple_service_locator/simple_service_locator.dart';
 
@@ -7,8 +8,13 @@ import 'package:simple_service_locator/simple_service_locator.dart';
 /// in [dispose]. Override [injectDependencies] to register state-local
 /// dependencies during [initState].
 mixin ScopedWidgetState<T extends StatefulWidget> on State<T> {
+  /// Name of the scope opened for this state.
+  ///
+  /// Override to provide deterministic or custom naming.
+  String get scopeName => '${runtimeType}Scope';
+
   /// Scope owned by this widget state.
-  late final DiScope scope = DiScope.open('${runtimeType}Scope');
+  late final DiScope scope;
 
   @override
   void initState() {
@@ -16,14 +22,17 @@ mixin ScopedWidgetState<T extends StatefulWidget> on State<T> {
     super.initState();
   }
 
+  /// Registers dependencies required by the widget state.
+  ///
+  /// Called from [initState] before `super.initState()`.
+  @mustCallSuper
+  void injectDependencies() {
+    scope = DiScope.open(scopeName);
+  }
+
   @override
   void dispose() {
     scope.close();
     super.dispose();
   }
-
-  /// Registers dependencies required by the widget state.
-  ///
-  /// Called from [initState] before `super.initState()`.
-  void injectDependencies() {}
 }